requirements.txt
# Core Framework
Django>=3.2,<4.2
djangorestframework>=3.14.0
markdown==3.5

# MongoDB Support
mongoengine>=0.27.0
pymongo>=4.3.0
dnspython>=2.3.0  # Required for MongoDB SRV connection strings

# Authentication & Security
pyjwt>=2.10.1
cryptography>=39.0
okta>=2.0.0

# Environment Management
python-dotenv>=0.20.0
django-environ>=0.11.2

# APIs & Web Services
requests>=2.31.0

# Static Files Management
whitenoise>=6.5.0

# Logging & Monitoring
loguru>=0.7.0
sentry-sdk>=1.30.0

# Data Handling & Analysis
pandas>=1.5.3
numpy>=1.24.0

# Deployment & Production
gunicorn>=21.2.0
django-prometheus>=2.3.1

# Task Queue
django-q2
redis>=4.6.0

# Development & Testing
django-debug-toolbar>=4.2.0
pytest-django>=4.5.2
django-extensions>=3.2.3

# Utilities
pathlib>=1.0.1

# Documentation
drf-yasg>=1.21.7
python-json-logger~=3.2.1